refactor: backgrounds 패키지 이름 수정, delete 코드 수정
refactor: backgrounds의 구문 오류 수정
