+++ TecheerPicture/src/main/java/com/techeerpicture/TecheerPicture/Background/PixelcutRequest.java
@@ -1,26 +1,33 @@
-package com.techeerpicture.TecheerPicture.Backgrounds;
+package com.techeerpicture.TecheerPicture.Background;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import io.swagger.v3.oas.annotations.media.Schema;
-import com.techeerpicture.TecheerPicture.Backgrounds.ImageTransform;
 
-public class BackgroundRequest {
-
+public class PixelcutRequest {
+    @JsonProperty("image_id") // JSON 키 이름을 API 요구사항에 맞게 설정
     private Long imageId;
+
+    @JsonProperty("image_url")
+    private String imageUrl;
+
+    @JsonProperty("image_transform")
     private ImageTransform imageTransform;
+
     private String scene;
+
     private String prompt;
+
+    @JsonProperty("negative_prompt")
     private String negativePrompt;
 
-    // 기본 생성자 추가
-    public BackgroundRequest() {}
+    public PixelcutRequest() {}
 
-    // 기존의 BackgroundService 주입 방식 제거
-    public BackgroundRequest(Long imageId, ImageTransform imageTransform, String scene, String prompt) {
+    public PixelcutRequest(Long imageId, String imageUrl, ImageTransform imageTransform, String scene, String prompt, String negativePrompt) {
         this.imageId = imageId;
+        this.imageUrl = imageUrl;
         this.imageTransform = imageTransform;
         this.scene = scene;
         this.prompt = prompt;
+        this.negativePrompt = negativePrompt;
     }
 
     // Getters and Setters
@@ -30,6 +37,14 @@ public class BackgroundRequest {
 
     public void setImageId(Long imageId) {
         this.imageId = imageId;
+    }
+
+    public String getImageUrl() {
+        return imageUrl;
+    }
+
+    public void setImageUrl(String imageUrl) {
+        this.imageUrl = imageUrl;
     }
 
     public ImageTransform getImageTransform() {
@@ -57,7 +72,7 @@ public class BackgroundRequest {
     }
 
     public String getNegativePrompt() {
-        return negativePrompt != null ? negativePrompt : "";
+        return negativePrompt;
     }
 
     public void setNegativePrompt(String negativePrompt) {
